+++ cliquet/tests/test_initialization.py
@@ -206,7 +206,7 @@ class StatsDConfigurationTest(unittest.TestCase):
         cliquet.initialize(self.config, '0.0.1')
         app = webtest.TestApp(self.config.make_wsgi_app())
         app.get('/v0/__heartbeat__', headers={'Authorization': 'Bearer abcde'})
-        mocked().count.assert_any_call('auth_type.FxA')
+        mocked().count.assert_any_call('authn_type.FxAOAuth')
 
 
 class RequestsConfigurationTest(unittest.TestCase):